rework multiblock-messages
Completely reworked the multiblock-messages. Now they use a new class with
its own thread. This way the data-transfer is completely async and it can
be added multiple messages, without any blocking of the statemashine.
Additionally with this setup, both sides can send data to each other
at the same time and streaming-messages are unaffected and can be send
at the same time too.